# Optimize internal clip_noise_rates and remove_noise_from_class functions