--- a/cleanlab/internal/util.py
+++ b/cleanlab/internal/util.py
@@ -21,7 +21,6 @@ Ancillary helper methods used internally throughout this package; mostly related
 import warnings
 from typing import Optional, Tuple, Union
 
-
 import numpy as np
 import pandas as pd